restructured all create-template redux stuff to have own reducer/actions etc, and updated content to include multiple reducers
