+++ scoresheet-app/src/App.jsx
@@ -5,8 +5,6 @@ import NavBar from './components/NavBar.jsx';
 
 // e.g. ReactBootstrap.Button, Navbar, Jumbotron, Table, Grid, collapse, Accordion, etc
 
-{/* <ReactBootstrap.collapse /> */}
-
 class App extends Component {
   render() {
     return (
