+++ scoresheet-app/src/redux/actions/create-templates.js
@@ -1,0 +1,8 @@
+import {CHANGE_TEMPLATE_NAME} from '../constants/create-templates';
+
+export function changeTemplateName(templateName) {
+  return {
+    type: CHANGE_TEMPLATE_NAME,
+    templateName
+  };
+}
