+++ scoresheet-app/src/redux/actions/pages.js
@@ -1,4 +1,5 @@
-import { CHANGE_PAGE, CHANGE_GAME_ID, CHANGE_TEMPLATE_NAME} from '../constants/pages';
+// import { CHANGE_PAGE, CHANGE_GAME_ID, CHANGE_TEMPLATE_NAME} from '../constants/pages';
+import { CHANGE_PAGE, CHANGE_GAME_ID} from '../constants/pages';
 
 export function changePage(page) {
   return {
@@ -14,9 +15,9 @@ export function changeGameID(gameid) {
   }
 }
 
-export function changeTemplateName(templateName) {
-  return {
-    type: CHANGE_TEMPLATE_NAME,
-    templateName
-  };
-}
+// export function changeTemplateName(templateName) {
+//   return {
+//     type: CHANGE_TEMPLATE_NAME,
+//     templateName
+//   };
+// }
