+++ scoresheet-app/src/redux/constants/create-templates.js
@@ -1,0 +1,1 @@
+export const CHANGE_TEMPLATE_NAME = 'CHANGE_TEMPLATE_NAME';
