+++ scoresheet-app/src/redux/reducers/pages.js
@@ -1,6 +1,8 @@
-import { CHANGE_PAGE, CHANGE_GAME_ID, CHANGE_TEMPLATE_NAME } from '../constants/pages';
+// import { CHANGE_PAGE, CHANGE_GAME_ID, CHANGE_TEMPLATE_NAME } from '../constants/pages';
+import { CHANGE_PAGE, CHANGE_GAME_ID} from '../constants/pages';
 
-export default (state = { page: 'home', gameid: '123', templateName: '' }, action) => {
+// export default (state = { page: 'home', gameid: '123', templateName: '' }, action) => {
+export default (state = { page: 'home', gameid: '123'}, action) => {
   switch (action.type) {
     case CHANGE_PAGE:
       return {
@@ -11,11 +13,6 @@ export default (state = { page: 'home', gameid: '123', templateName: '' }, actio
       return {
         ...state,
         gameid: action.gameid
-      }
-    case CHANGE_TEMPLATE_NAME:
-      return {
-        ...state,
-        templateName: action.templateName
       }
     default:
       return state;
